Trait repository::config::ConfigContract

pub trait ConfigContract {
    fn has(&self, key: &str) -> bool;
    fn get(&self, key: &str) -> Option<&str>;
    fn set(&mut self, key: &str, value: &str);
    fn all(&self) -> &HashMap<String, String>;
}

A contract for configuration management.

Required Methods

fn has(&self, key: &str) -> bool

Checks if a given key exists in the configuration.

Examples

use repository::config::{ConfigContract, Repository};
use std::collections::HashMap;

let config_items = [("foo".to_string(), "bar".to_string())].iter().cloned().collect::<HashMap<_,_>>();
let config = Repository::new(config_items);
assert_eq!(config.has("foo"), true);
assert_eq!(config.has("missing"), false);

fn get(&self, key: &str) -> Option<&str>

Gets the value associated with the given key in the configuration.

Examples

use repository::config::{ConfigContract, Repository};
use std::collections::HashMap;

let config_items = [("foo".to_string(), "bar".to_string())].iter().cloned().collect::<HashMap<_,_>>();
let config = Repository::new(config_items);
assert_eq!(config.get("foo"), Some("bar"));
assert_eq!(config.get("missing"), None);

fn set(&mut self, key: &str, value: &str)

Sets the value associated with the given key in the configuration.

Examples

use repository::config::{ConfigContract, Repository};
use std::collections::HashMap;

let mut config_items = [("foo".to_string(), "bar".to_string())].iter().cloned().collect::<HashMap<_,_>>();
let mut config = Repository::new(config_items);
config.set("foo", "baz");
assert_eq!(config.get("foo"), Some("baz"));

fn all(&self) -> &HashMap<String, String>

Gets all configuration items as a reference to a hashmap.

Examples

use repository::config::{ConfigContract, Repository};
use std::collections::HashMap;

let config_items = [("foo".to_string(), "bar".to_string())].iter().cloned().collect::<HashMap<_,_>>();
let config = Repository::new(config_items);
assert_eq!(config.all().get("foo"), Some(&"bar".to_string()));